Begin writing a syntax highlighter for jira markup
I'm not sure of the actual grammar for jira formatting, and the
documentation is quite sparse; the documentation really only includes
simple examples and doesn't outline any of the complexities such as
what happens when other text surrounds some of the symbols.

The rules are implemented based on what I've observed at the current
time. Most of this behaviour is subject to change at any point (or
may already be different to what I've implemented).